feat: Xbox controllers

+++ maxine-lib/src/lib.rs
@@ -1,15 +1,17 @@
+#![allow(mixed_script_confusables)]
+
 pub mod transforms {
-    use std::f32::consts::FRAC_1_PI;
+    use std::f64::consts::FRAC_1_PI;
 
     use nalgebra::Isometry2;
 
     /// TODO: move this to Arfur.
     /// TODO: review speed/benchmark and optimise
-    pub fn deadband(x: f32) -> f32 {
-        const CUTOFF: f32 = 0.1;
-        const WEIGHT: f32 = 0.2;
+    pub fn deadband(x: f64) -> f64 {
+        const CUTOFF: f64 = 0.1;
+        const WEIGHT: f64 = 0.2;
 
-        fn cubic(x: f32, weight: f32) -> f32 {
+        fn cubic(x: f64, weight: f64) -> f64 {
             weight * x * x * x + (1. - weight) * x
         }
 
@@ -22,7 +24,7 @@ pub mod transforms {
 
     /// Given an isometry, calculate the rotational and driving values
     /// (between -1 and 1) for a swerve module.
-    pub fn swerve(isometry: Isometry2<f32>) -> (f32, f32) {
+    pub fn swerve(isometry: Isometry2<f64>) -> (f64, f64) {
         // Extract some information from the isometry for ease-of-use.
         let x = isometry.translation.x;
         let y = isometry.translation.y;
